# Conflicts:
#	src/data/rt-handlers.js

+++ test/e2e/helpers/sandbox.js
@@ -1,16 +1,21 @@
+import Request from 'backendless-request'
 import { createClient } from 'backendless-console-sdk'
-import { TablesAPI } from './tables'
-import { wait } from './promise'
-import * as Utils from './utils'
-// import Backendless from '../../../src/backendless'
+
+import '../../helpers/global'
+import Utils from '../../helpers/utils'
+
 const Backendless = require('../../../lib')
+
+import { TablesAPI } from './tables'
 
 const API_SERVER = process.env.API_SERVER || 'http://localhost:9000'
 const CONSOLE_SERVER = process.env.CONSOLE_SERVER || 'http://localhost:3000'
 const TEST_USER_EMAIL = process.env.TEST_USER_EMAIL || 'foo@foo.com'
 const TEST_USER_PASSWORD = process.env.TEST_USER_PASSWORD || 'secret'
-const DESTROY_ALL_PREV_APPS = process.env.DESTROY_ALL_PREV_APPS === 'true'
+const TEST_VERBOSE = process.env.TEST_VERBOSE === 'true'
+const DESTROY_ALL_PREV_APPS = process.env.DESTROY_ALL_PREV_APPS !== 'false'
 const DESTROY_APP_AFTER_TEST = process.env.DESTROY_APP_AFTER_TEST !== 'false'
+const PERSISTED_APP_ID = process.env.PERSISTED_APP_ID
 
 const TEST_APP_NAME_PATTERN = /^test_.{32}$/
 
@@ -19,9 +24,11 @@ const persistedLocalDev = () => ({
   pwd  : TEST_USER_PASSWORD
 })
 
-const generateApp = () => ({
-  name: `test_${Utils.uid()}`
-})
+const log = (...args) => {
+  if (TEST_VERBOSE) {
+    console.log(...args)
+  }
+}
 
 const destroyAllTestApps = (() => {
   let prevRequest = null
@@ -31,12 +38,18 @@ const destroyAllTestApps = (() => {
       return prevRequest
     }
 
+    log('destroy all previous test apps')
+
     return prevRequest = Promise.resolve()
       .then(async () => {
         const apps = await api.apps.getApps()
 
-        await Promise.all(apps.map(app => {
+        log(`loaded ${apps.length} apps`)
+
+        return Promise.all(apps.map(app => {
           if (TEST_APP_NAME_PATTERN.test(app.name)) {
+            log(`destroy app: ${app.id} (${app.name})`)
+
             return api.apps.deleteApp(app.id)
           }
         }))
@@ -45,19 +58,35 @@ const destroyAllTestApps = (() => {
 })()
 
 const createDestroyer = sandbox => async () => {
-  if (DESTROY_APP_AFTER_TEST) {
+  if (DESTROY_APP_AFTER_TEST && !PERSISTED_APP_ID) {
+    log(`destroy after tests application: ${sandbox.app.id} (${sandbox.app.name})`)
+
     await sandbox.api.apps.deleteApp(sandbox.app.id)
   }
 }
 
+const provideApp = api => {
+  if (PERSISTED_APP_ID) {
+    return {
+      id: PERSISTED_APP_ID
+    }
+  }
+
+  return api.apps.createApp({ appName: `test_${Utils.uid()}`, refCode: null })
+}
+
 const createSandbox = async api => {
-  const app = generateApp()
+  log('create sandbox')
+
+  const app = {}
   const dev = persistedLocalDev()
 
   const sandbox = { app, api, dev }
   sandbox.destroy = createDestroyer(sandbox)
 
   const user = await api.user.login(dev.email, dev.pwd)
+
+  log('logged in user: ', dev.email)
 
   dev.id = user.id
   dev.name = user.name
@@ -67,13 +96,17 @@ const createSandbox = async api => {
     await destroyAllTestApps(api)
   }
 
-  const createApp = await api.apps.createApp({ appName: app.name, refCode: null })
+  const createApp = await provideApp(api)
 
   Object.assign(app, createApp)
+
+  log(`load app settings for app: ${sandbox.app.id} (${sandbox.app.name})`)
 
   const appSettings = await api.settings.getAppSettings(app.id)
 
   Object.assign(app, appSettings)
+
+  log('created sandbox', sandbox)
 
   return sandbox
 }
@@ -87,7 +120,7 @@ const waitUntilAppIsConfigured = async app => {
   } catch (e) {
     console.log('App is not ready yet', e)
 
-    await wait(5000)
+    await Utils.wait(5000)
 
     await waitUntilAppIsConfigured(app)
   }
@@ -95,12 +128,54 @@ const waitUntilAppIsConfigured = async app => {
   app.ready = true
 }
 
+let mockRequests = []
+const nativeRequestSend = Request.send
+
+Request.send = function fakeRequestSend(path, method, headers, body, encoding) {
+  const mockRequest = mockRequests.shift()
+
+  if (!mockRequest) {
+    return nativeRequestSend.call(Request, path, method, headers, body, encoding)
+  }
+
+  Object.assign(mockRequest, { path, method, headers, body, encoding })
+
+  try {
+    mockRequest.body = JSON.parse(body)
+  } catch (e) {
+    //
+  }
+
+  const responseFn = mockRequest.responseFn || Promise.resolve()
+
+  return responseFn
+    .then(response => {
+      return Object.assign({ status: 200, body: undefined, headers: {} }, response)
+    })
+}
+
+function prepareMockRequest(responseFn) {
+  const mockRequest = {
+    responseFn
+  }
+
+  mockRequests.push(mockRequest)
+
+  return mockRequest
+}
+
 const createSandboxFor = each => () => {
   const beforeHook = each ? beforeEach : before
   const afterHook = each ? afterEach : after
 
+  const timeout = 10 * 60000 // 10 minutes
+
+  beforeEach(() => {
+    mockRequests = []
+  })
+
   beforeHook(function() {
-    this.timeout(120000)
+    this.timeout(timeout + (timeout * 0.2))
     this.consoleApi = createClient(CONSOLE_SERVER)
     this.tablesAPI = new TablesAPI(this)
 
@@ -110,10 +185,14 @@ const createSandboxFor = each => () => {
         this.dev = sandbox.dev
         this.app = sandbox.app
 
+        Backendless.debugMode = TEST_VERBOSE
         Backendless.serverURL = API_SERVER
         Backendless.initApp(this.app.id, this.app.apiKeysMap.JS)
       })
-      .then(() => Promise.race([waitUntilAppIsConfigured(this.app), wait(120000)]))
+      .then(() => Promise.race([
+        waitUntilAppIsConfigured(this.app),
+        Utils.wait(timeout)
+      ]))
       .then(() => {
         if (!this.app.ready) {
           throw new Error('App was created with error!')
@@ -128,8 +207,13 @@ const createSandboxFor = each => () => {
   })
 }
 
+export {
+  Utils
+}
+
 export default {
   Backendless,
+  prepareMockRequest,
   forTest : createSandboxFor(true),
   forSuite: createSandboxFor(false)
 }
